nostr: split backend.go into focused files
backend.go grew to 965 lines mixing cryptography, media handling, and DM
construction with core lifecycle logic. This made it harder to navigate
and reason about individual concerns.

Extract pure AES-GCM decryption helpers into crypto.go, file/media
download and URL rewriting into media.go, and DM building/sending with
reaction logic into dm.go. backend.go drops to 616 lines and each new
file stays under 140 lines with a single clear responsibility.
