+++ nostr/backend.go
@@ -3,26 +3,15 @@ package nostr
 
 import (
 	"context"
-	"crypto/aes"
-	"crypto/cipher"
-	"crypto/sha256"
-	"encoding/hex"
 	"encoding/json"
-	"errors"
 	"fmt"
 	"log/slog"
-	"mime"
-	"os"
-	"path/filepath"
-	"regexp"
-	"strconv"
 	"strings"
 	"sync"
 	"time"
 
 	gonostr "fiatjaf.com/nostr"
 	"fiatjaf.com/nostr/keyer"
-	"fiatjaf.com/nostr/nip17"
 	"fiatjaf.com/nostr/nip59"
 	"github.com/pinpox/opencrow/backend"
 )
@@ -484,127 +473,6 @@ func (b *Backend) pruneSeen() {
 	b.seenRumorsMu.Unlock()
 }
 
-func (b *Backend) sendDM(ctx context.Context, kr gonostr.Keyer, pool *gonostr.Pool, recipientPK gonostr.PubKey, text string, extraTags gonostr.Tags) string {
-	rumor, err := b.buildDMRumor(ctx, kr, recipientPK, text, extraTags)
-	if err != nil {
-		slog.Error("nostr: failed to build DM rumor", "recipient", recipientPK.Hex(), "error", err)
-
-		return ""
-	}
-
-	toUs, err := nip59.GiftWrap(rumor, rumor.PubKey,
-		func(s string) (string, error) { return kr.Encrypt(ctx, s, rumor.PubKey) },
-		func(e *gonostr.Event) error { return kr.SignEvent(ctx, e) },
-		nil,
-	)
-	if err != nil {
-		slog.Error("nostr: failed to gift-wrap DM (toUs)", "recipient", recipientPK.Hex(), "error", err)
-
-		return ""
-	}
-
-	toThem, err := nip59.GiftWrap(rumor, recipientPK,
-		func(s string) (string, error) { return kr.Encrypt(ctx, s, recipientPK) },
-		func(e *gonostr.Event) error { return kr.SignEvent(ctx, e) },
-		nil,
-	)
-	if err != nil {
-		slog.Error("nostr: failed to gift-wrap DM (toThem)", "recipient", recipientPK.Hex(), "error", err)
-
-		return ""
-	}
-
-	b.publishDMGiftWraps(ctx, pool, toUs, toThem, recipientPK)
-
-	return rumor.ID.Hex()
-}
-
-// buildDMRumor constructs and returns a kind 14 rumor event for a DM.
-func (b *Backend) buildDMRumor(ctx context.Context, kr gonostr.Keyer, recipientPK gonostr.PubKey, text string, extraTags gonostr.Tags) (gonostr.Event, error) {
-	ourPubkey, err := kr.GetPublicKey(ctx)
-	if err != nil {
-		return gonostr.Event{}, fmt.Errorf("getting public key: %w", err)
-	}
-
-	tags := make(gonostr.Tags, 0, len(extraTags)+1)
-	tags = append(tags, extraTags...)
-	tags = append(tags, gonostr.Tag{"p", recipientPK.Hex()})
-
-	rumor := gonostr.Event{
-		Kind:      gonostr.KindDirectMessage,
-		Content:   text,
-		Tags:      tags,
-		CreatedAt: gonostr.Now(),
-		PubKey:    ourPubkey,
-	}
-	rumor.ID = rumor.GetID()
-
-	return rumor, nil
-}
-
-// publishDMGiftWraps publishes the two gift-wrap copies to the appropriate relays.
-func (b *Backend) publishDMGiftWraps(ctx context.Context, pool *gonostr.Pool, toUs, toThem gonostr.Event, recipientPK gonostr.PubKey) {
-	// Publish our copy to our DM relays.
-	b.publishToRelays(toUs, b.cfg.DMRelays, "DM toUs")
-
-	theirRelays := b.recipientRelays(ctx, recipientPK, pool)
-
-	slog.Debug("nostr: sending DM", "recipient", recipientPK.Hex(), "their_relays", theirRelays)
-
-	b.publishToRelays(toThem, theirRelays, "DM toThem")
-}
-
-// sendReaction sends a NIP-25 kind 7 reaction gift-wrapped via NIP-59 to
-// acknowledge receipt of an incoming DM. The reaction references the original
-// rumor by its event ID so the sender's client can attach it to the right message.
-func (b *Backend) sendReaction(ctx context.Context, rumorID gonostr.ID, recipientPK gonostr.PubKey, rumorKind gonostr.Kind) {
-	rumor := gonostr.Event{
-		Kind:      gonostr.KindReaction,
-		Content:   "👍",
-		CreatedAt: gonostr.Now(),
-		PubKey:    b.keys.PK,
-		Tags: gonostr.Tags{
-			{"e", rumorID.Hex()},
-			{"p", recipientPK.Hex()},
-			{"k", strconv.Itoa(int(rumorKind))},
-		},
-	}
-	rumor.ID = rumor.GetID()
-
-	// Gift-wrap to recipient.
-	toThem, err := nip59.GiftWrap(
-		rumor,
-		recipientPK,
-		func(s string) (string, error) { return b.kr.Encrypt(ctx, s, recipientPK) },
-		func(e *gonostr.Event) error { return b.kr.SignEvent(ctx, e) },
-		nil,
-	)
-	if err != nil {
-		slog.Warn("nostr: failed to gift-wrap reaction", "recipient", recipientPK.Hex(), "error", err)
-
-		return
-	}
-
-	b.publishToRelays(toThem, b.recipientRelays(ctx, recipientPK, b.pool), "reaction")
-}
-
-// recipientRelays resolves the recipient's DM relays via NIP-17 discovery,
-// falling back to our own DM relays when the recipient has no list published.
-func (b *Backend) recipientRelays(ctx context.Context, recipientPK gonostr.PubKey, pool *gonostr.Pool) []string {
-	relays := nip17.GetDMRelays(ctx, recipientPK, pool, b.cfg.Relays)
-	if len(relays) == 0 {
-		return b.cfg.DMRelays
-	}
-
-	return relays
-}
-
-// publishToRelays enqueues an event for publishing to the given relays.
-// The actual publishing happens asynchronously via the queue's drain loop.
-func (b *Backend) publishToRelays(evt gonostr.Event, relays []string, label string) {
-	b.pubQueue.enqueue(evt, relays, label)
-}
-
 // processGiftWrap unwraps a kind 1059 event and dispatches to the handler.
 func (b *Backend) processGiftWrap(ctx, pubCtx context.Context, evt *gonostr.Event) {
 	if evt == nil {
@@ -732,223 +600,6 @@ func (b *Backend) claimConversation(senderHex string) bool {
 	}
 
 	return true
-}
-
-// KindFileMessage is the NIP-17 kind for file messages (not yet in go-nostr).
-const KindFileMessage gonostr.Kind = 15
-
-var mediaURLRe = regexp.MustCompile(`(?i)https?://\S+\.(?:png|jpg|jpeg|gif|webp|pdf|mp3|mp4|wav|ogg|svg|bmp|tiff|zip)(?:\?\S*)?`)
-
-// handleFileMessage processes a NIP-17 kind 15 file message. The rumor's
-// content is the file URL and tags carry metadata (file-type, dimensions, etc.).
-// The file is downloaded and the message text is rewritten to reference the
-// local path so pi can read it with its tools.
-func (b *Backend) handleFileMessage(ctx context.Context, rumor gonostr.Event, conversationID string) string {
-	fileURL := strings.TrimSpace(rumor.Content)
-	if fileURL == "" {
-		slog.Warn("nostr: kind 15 file message with empty content", "sender", conversationID)
-
-		return "[User sent a file message with no URL]"
-	}
-
-	mimeType := tagValue(rumor.Tags, "file-type")
-
-	slog.Info("nostr: downloading file message", "url", fileURL, "mime", mimeType, "sender", conversationID)
-
-	localPath, err := downloadURL(ctx, fileURL, b.cfg.SessionBaseDir, conversationID)
-	if err != nil {
-		slog.Warn("nostr: failed to download file message", "url", fileURL, "error", err)
-
-		return fmt.Sprintf("[User sent a file but download failed: %s]", fileURL)
-	}
-
-	// Decrypt if the file was encrypted (NIP-17 kind 15 AES-GCM).
-	if algo := tagValue(rumor.Tags, "encryption-algorithm"); algo != "" {
-		if err := decryptFileInPlace(localPath, rumor.Tags); err != nil {
-			slog.Warn("nostr: failed to decrypt file message", "url", fileURL, "error", err)
-			os.Remove(localPath)
-
-			return "[User sent an encrypted file but decryption failed]"
-		}
-	}
-
-	// If the downloaded file has no extension, try to add one from the mime type.
-	localPath = maybeAddExtension(localPath, mimeType)
-
-	desc := descFromMIME(mimeType)
-
-	return fmt.Sprintf("[User sent a %s: %s]\nUse the read tool to view it.", desc, localPath)
-}
-
-// decryptFileInPlace reads the file at path, decrypts it using AES-GCM with
-// the key and nonce from the rumor tags, verifies the SHA-256 hash against
-// the "ox" tag (pre-encryption hash), and writes the plaintext back.
-func decryptFileInPlace(filePath string, tags gonostr.Tags) error {
-	algo := tagValue(tags, "encryption-algorithm")
-	if algo != "aes-gcm" {
-		return fmt.Errorf("unsupported encryption algorithm: %q", algo)
-	}
-
-	key, nonce, err := parseDecryptionParams(tags)
-	if err != nil {
-		return err
-	}
-
-	ciphertext, err := os.ReadFile(filePath)
-	if err != nil {
-		return fmt.Errorf("reading encrypted file: %w", err)
-	}
-
-	plaintext, err := decryptAESGCM(key, nonce, ciphertext)
-	if err != nil {
-		return err
-	}
-
-	// Verify against the pre-encryption hash if provided.
-	if oxHex := tagValue(tags, "ox"); oxHex != "" {
-		hash := sha256.Sum256(plaintext)
-
-		if hex.EncodeToString(hash[:]) != oxHex {
-			return errors.New("SHA-256 mismatch after decryption")
-		}
-	}
-
-	if err := os.WriteFile(filePath, plaintext, 0o600); err != nil {
-		return fmt.Errorf("writing decrypted file: %w", err)
-	}
-
-	return nil
-}
-
-// parseDecryptionParams extracts and decodes the AES key and nonce from tags.
-func parseDecryptionParams(tags gonostr.Tags) ([]byte, []byte, error) {
-	keyHex := tagValue(tags, "decryption-key")
-	nonceHex := tagValue(tags, "decryption-nonce")
-
-	if keyHex == "" || nonceHex == "" {
-		return nil, nil, errors.New("missing decryption-key or decryption-nonce tags")
-	}
-
-	key, err := hex.DecodeString(keyHex)
-	if err != nil {
-		return nil, nil, fmt.Errorf("decoding decryption key: %w", err)
-	}
-
-	nonce, err := hex.DecodeString(nonceHex)
-	if err != nil {
-		return nil, nil, fmt.Errorf("decoding decryption nonce: %w", err)
-	}
-
-	return key, nonce, nil
-}
-
-// decryptAESGCM decrypts ciphertext using AES-GCM with the given key and nonce.
-func decryptAESGCM(key, nonce, ciphertext []byte) ([]byte, error) {
-	block, err := aes.NewCipher(key)
-	if err != nil {
-		return nil, fmt.Errorf("creating AES cipher: %w", err)
-	}
-
-	gcm, err := cipher.NewGCMWithNonceSize(block, len(nonce))
-	if err != nil {
-		return nil, fmt.Errorf("creating GCM: %w", err)
-	}
-
-	plaintext, err := gcm.Open(nil, nonce, ciphertext, nil)
-	if err != nil {
-		return nil, fmt.Errorf("AES-GCM decryption: %w", err)
-	}
-
-	return plaintext, nil
-}
-
-// tagValue returns the value of the first tag with the given key, or "".
-func tagValue(tags gonostr.Tags, key string) string {
-	for _, tag := range tags {
-		if len(tag) >= 2 && tag[0] == key {
-			return tag[1]
-		}
-	}
-
-	return ""
-}
-
-// maybeAddExtension renames a file to include a proper extension when the
-// downloaded filename has none and the mime type is known.
-func maybeAddExtension(path string, mimeType string) string {
-	if mimeType == "" || filepath.Ext(path) != "" {
-		return path
-	}
-
-	ext := mimeToExt(mimeType)
-	if ext == "" {
-		return path
-	}
-
-	newPath := path + ext
-	if err := os.Rename(path, newPath); err != nil {
-		slog.Warn("nostr: failed to rename file with extension", "from", path, "to", newPath, "error", err)
-
-		return path
-	}
-
-	return newPath
-}
-
-// descFromMIME returns a human-readable file type description for the
-// attachment message text.
-func descFromMIME(mimeType string) string {
-	switch {
-	case strings.HasPrefix(mimeType, "image/"):
-		return "image"
-	case strings.HasPrefix(mimeType, "audio/"):
-		return "audio"
-	case strings.HasPrefix(mimeType, "video/"):
-		return "video"
-	default:
-		return "file"
-	}
-}
-
-// mimeToExt returns a file extension for a MIME type using the system MIME
-// database (via mime.ExtensionsByType). Returns "" for unknown types.
-func mimeToExt(mimeType string) string {
-	exts, err := mime.ExtensionsByType(mimeType)
-	if err != nil || len(exts) == 0 {
-		return ""
-	}
-
-	return exts[0]
-}
-
-// rewriteMediaURLs finds media URLs in text, downloads them, and rewrites the
-// text with the local file path in the standard attachment format.
-func (b *Backend) rewriteMediaURLs(ctx context.Context, text, conversationID string) string {
-	urls := mediaURLRe.FindAllString(text, -1)
-	if len(urls) == 0 {
-		return text
-	}
-
-	processed := make(map[string]bool, len(urls))
-	for _, rawURL := range urls {
-		if processed[rawURL] {
-			continue
-		}
-
-		localPath, err := downloadURL(ctx, rawURL, b.cfg.SessionBaseDir, conversationID)
-		if err != nil {
-			slog.Warn("nostr: failed to download media URL", "url", rawURL, "error", err)
-
-			continue
-		}
-
-		slog.Debug("nostr: downloaded media URL", "url", rawURL, "localPath", localPath)
-		processed[rawURL] = true
-		replacement := fmt.Sprintf("[User sent a file (no caption): %s]\nUse the read tool to view it.", localPath)
-		text = strings.ReplaceAll(text, rawURL, replacement)
-	}
-
-	return text
 }
 
 // rumorReplyTarget returns the value of the first "e" tag in the rumor, or ""
